app/backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-rest-auth==0.9.5
django-allauth==0.54.0
django-cors-headers==3.8.0
django-heroku==0.3.1
mysqlclient==2.1.1
gunicorn==20.1.0
